Fix u32 overflow allowing OOB read in TVOL texture extraction
offset+size was computed as u32 addition; a crafted size near
UINT32_MAX wraps the sum below raw_size, passing the bounds check
and making to_write the full attacker-controlled size, so SaveFile
copies far past the end of the loaded buffer. Compute the sum in
64-bit before comparing against raw_size.

+++ project/src/lib-tvol.c
@@ -52,7 +52,7 @@ enumError ExtractTVOLArchive (ccp arg, ccp basedir, uint depth)
 		if (size == 0 || offset >= raw_size)
 			continue;
 
-		const u32 to_write = offset + size <= raw_size ? size : (u32)(raw_size - offset);
+		const u32 to_write = (u64)offset + size <= raw_size ? size : (u32)(raw_size - offset);
 
 		// Name is stored at offset as null-terminated string (up to 48 bytes)
 		char name[64] = "";
